rename lesoson input to create-lesson to be more specific
